complete tests refactor
